Fix dynamic image
