+++ packages/mobile/src/components/core/index.ts
@@ -1,7 +1,7 @@
 export * from './AnimatedButton'
 export * from './Button'
 export * from './CardList'
-export * from './ImageLoader'
+export * from './DynamicImage'
 export * from './EmptyTile'
 export * from './FormTextInput'
 export * from './FormImageInput'
