+++ packages/mobile/src/components/lineup-tile/types.ts
@@ -13,7 +13,7 @@ import type {
 
 import type { GestureResponderHandler } from 'app/types/gesture'
 
-import type { ImageLoaderProps, TileProps } from '../core'
+import type { DynamicImageProps, TileProps } from '../core'
 
 export type LineupItemProps = {
   /** Index of tile in lineup */
@@ -63,7 +63,7 @@ export type LineupTileProps = Omit<LineupItemProps, 'togglePlay'> & {
   id: ID
 
   /** Render function for the image */
-  renderImage: (props: ImageLoaderProps) => ReactNode
+  renderImage: (props: DynamicImageProps) => ReactNode
 
   /** The item (track or collection) */
   item: Track | Collection
